legger til mulighet for å sende med et array med onClickHandlers til <Table/> som gjør at du kan legge click-handlers på tablerows
